requirements.txt
flask==3.0.0
pymongo
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
langchain-experimental
apscheduler
bcrypt
requests
pillow
pypdf
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pinecone==3.1.0
